+++ src/Contexts/auth.js
@@ -57,7 +57,7 @@ export default function AuthContextProvider({ children }) {
 
         setUser(user);
         setAccessToken(accessToken);
-        navigation.navigate("InicioRoutes");
+        navigation.navigate("SignedIn");
       } catch (err) {
         console.log(err);
         showMessage({
