+++ src/Pages/Success/index.js
@@ -18,7 +18,7 @@ export default function Success({ navigation }) {
           marginLeft={0}
           background={"white"}
           size={18}
-          onPress={() => navigation.navigate("Home")}
+          onPress={() => navigation.navigate("SignedIn")}
         />
       </ButtonContainer>
     </Container>
